Revert "Rename go.yml to .github/workflows/go.yml (#14)"
This reverts commit f126f655867778806cdadfe092d3ee50ecb51c97.
