+++ cmd/ufs/ufs.go
@@ -9,8 +9,8 @@ import (
 	"log"
 	"net"
 
-	"harvey-os.org/ninep/protocol"
-	"harvey-os.org/ninep/ufs"
+	"harvey-os.org/internal/ufs"
+	"harvey-os.org/pkg/ninep/protocol"
 )
 
 var (
