+++ internal/tmpfs/archive.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"harvey-os.org/ninep/protocol"
+	"harvey-os.org/pkg/ninep/protocol"
 )
 
 // Archive contains the directories and files from a decompressed archive
