+++ pkg/ninep/protocol/genout.go
